+++ src/controllers/LoginController.js
@@ -1,6 +1,6 @@
 const Aluno = require("../models/Aluno")
 const { sign } = require('jsonwebtoken')
-
+const { compare, hash } = require("bcryptjs")
 class LoginController {
 
     async login(req, res) {
@@ -22,6 +22,11 @@ class LoginController {
 
             if (!aluno) {
                 return res.status(404).json({ error: 'Nenhum aluno corresponde a email e senha fornecidos!' })
+            }
+            const hashSenha = await compare(password, aluno.password)
+
+            if(hashSenha === false) {
+                return res.status(403).json({mensagem: 'Aluno não encontrado'})
             }
 
             const payload = { sub: aluno.id, email: aluno.email, nome: aluno.nome }
