+++ src/models/Aluno.js
@@ -1,5 +1,6 @@
 const { DataTypes } = require('sequelize')
 const { connection } = require('../database/connection')
+const { hash }  = require('bcryptjs')
 
 const Aluno = connection.define('alunos', {
     email:{
@@ -19,6 +20,12 @@ const Aluno = connection.define('alunos', {
     }
 })
 
+// hooks
+Aluno.beforeSave(async (user) =>{ //before save vai utilizar antes de cadastrar um usuario 
+    '' //recupera o usuario que sera criado
+       user.password = await hash(user.password, 8)
+       return user
+   })
 module.exports = Aluno
